requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
filelock==3.12.2
Flask==2.3.2
Flask-Reuploaded==1.3.0
Flask-Uploads==0.2.1
Flask-WTF==1.1.1
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.0
Pillow==9.5.0
requests==2.31.0
sympy==1.12
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.3.6
WTForms==3.0.1
zipp==3.15.0
